Added a hacky silk touch upgrade
It works as it should, however due to the fact that the silk touch
info of the block is "protected", it has to exploit the fact that
objects in the same package can access the protected members of
classes. Therefore, the SilkTouchHarvest helper resides in the
"net.minecraft.block" package and has access to block's silk touch info.