Implement a speed upgrade and energy addition for upgrades
The upgrades now consume more energy, currently up to about 6k with
all the upgrades and even more for the ores that require stronger
tools. That required bumping the max energy receiveing rate some
more. Maybe this could also increase the max energy amount stored in
the machine in the future.