Refactor "resources left" counter to work from the start
There's now a field that is set to true that makes the kuarry recalculate
the resource left count on the next update. It's set to true by
default, so when it starts it calculates the amount of resources
needed, as it should. Apparently it didn't before, for some reason
"onLoad" doesn't work well for that.