Implement leveling the kuarry up
There are currently three levels levels, 0, 1 and 2 (1, 2 and 3 for
the user). The levels add more energy and upgrade slots for the
upcoming upgrades, since there might not be enough space when they're
implemented. While the third level gives 10 upgrade slots, which
really won't have anything to fill them for a while, the system for
expanding the slot amount easily is implemented, so when some upgrades
DO come, there will already be place for them.

The upgrades are done by the two new upgrade items. They have slightly
different colors and names, inspired by the upgrades in
Thermal Expansion.

The block model now selects the the appropriate data for the
blockstate in getActualState, getting it from the tile entity. Same
thing happens for the ItemStack name and the custom mesh definition.

For reasons unknown, the "inventory" variation has to stay in the
blockstate, because oterwise the game throws an exception about it not
beign there, albeit everything else working normally.