Implement XP fluid colection & make incompat upgrades an array
The XP upgrade uses the new XP collection component that collects two
kinds of XP fluids (from te and from enderio) into the inner fluid
container. They both give 20mB of the fluid as this seems to the
amount that TE uses and enderio seems to use about the same amount.

The XP collection upgrade is incompatible with the silk touch upgrade,
as it doesn't make sense to give XP for blocks that are not actually
mined normally.

Closes #6